% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglShiny.R
\name{center}
\alias{center}
\title{place the specified representation into the center of ngl's stage}
\usage{
center(session, htmlContainer, selectionString)
}
\arguments{
\item{session}{a Shiny server session object.}

\item{htmlContainer}{character}

\item{selectionString}{character using ngl's selection language}
}
\description{
place the specified representation into the center of ngl's stage
}
\examples{
\dontrun{
center(session, htmlContainer="nglShiny_1s5l", "OEC AND :A")
}

}
