% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglShiny.R
\name{fit}
\alias{fit}
\title{Set zoom and center so that the current model nicely fills the display.}
\usage{
fit(session)
}
\arguments{
\item{session}{a Shiny server session object.}
}
\description{
Set zoom and center so that the current model nicely fills the display.
}
\examples{
\dontrun{
  fit(session)
}

}
