% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglShiny.R
\name{nglShiny}
\alias{nglShiny}
\title{nglShiny
nglShiny}
\usage{
nglShiny(options, width = NULL, height = NULL, elementId)
}
\arguments{
\item{width}{integer  initial width of the widget.}

\item{height}{integer initial height of the widget.}

\item{elementId}{string the DOM id into which the widget is rendered, default NULL is best.}
}
\value{
a reference to an htmlwidget.
}
\description{
This widget wraps ngl.js a WebGL protein viewer
https://github.com/arose/ngl
nglShiny

This widget wraps cytoscape.js, a full-featured Javsscript network library for visualization and analysis.
}
