% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglShiny-class.R, R/nglShiny.R
\name{nglShinyOutput}
\alias{nglShinyOutput}
\title{Standard shiny ui rendering construct}
\usage{
nglShinyOutput(outputId, width = "100\%", height = "600px")

nglShinyOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{the name of the DOM element to create.}

\item{width}{integer  optional initial width of the widget.}

\item{height}{integer optional initial height of the widget.}
}
\value{
a reference to an htmlwidget

a reference to an htmlwidget
}
\description{
Standard shiny ui rendering construct

Standard shiny ui rendering construct
}
\examples{
\dontrun{
  mainPanel(nglShinyOutput('nglShiny'), width=10)
}

\dontrun{
  mainPanel(nglShinyOutput('nglShiny'), width=10)
}

}
