% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglShiny-class.R, R/nglShiny.R
\name{renderNglShiny}
\alias{renderNglShiny}
\title{More shiny plumbing -  an nglShiny wrapper for htmlwidget standard rendering operation}
\usage{
renderNglShiny(expr, env = parent.frame(), quoted = FALSE)

renderNglShiny(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates an HTML widget.}

\item{env}{environment in which to evaluate expr.}

\item{quoted}{logical specifies whether expr is quoted ("useuful if you want to save an expression in a variable").}
}
\value{
not sure

not sure
}
\description{
More shiny plumbing -  an nglShiny wrapper for htmlwidget standard rendering operation

More shiny plumbing -  an nglShiny wrapper for htmlwidget standard rendering operation
}
