% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglShiny.R
\name{spin}
\alias{spin}
\title{set the new spin state}
\usage{
spin(session, newState)
}
\arguments{
\item{session}{a Shiny server session object.}

\item{newState}{logical, either TRUE or FALSE}
}
\description{
set the new spin state
}
