% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglShiny.R
\name{showSelection}
\alias{showSelection}
\title{Using the specified representation and colorScheme, display the portion of selection}
\usage{
showSelection(session, representation, selection, colorScheme = "residueIndex")
}
\arguments{
\item{session}{a Shiny server session object.}

\item{representation}{todo}

\item{selection}{todo}

\item{colorScheme}{todo}
}
\description{
Using the specified representation and colorScheme, display the portion of selection
}
\examples{
\dontrun{
  showSelection(session, "cartoon", "helix", "residueIndex")
}

}
